+++ apps/betterangels/src/app/(private-screens)/add-note/Team/index.tsx
@@ -1,17 +1,13 @@
 import {
-  enumDisplaySelahTeam,
   SelahTeamEnum,
+  enumDisplaySelahTeam,
   useSnackbar,
   useUpdateNoteMutation,
 } from '@monorepo/expo/betterangels';
 import { Spacings } from '@monorepo/expo/shared/static';
-import { Picker } from '@monorepo/expo/shared/ui-components';
+import { SingleSelect } from '@monorepo/expo/shared/ui-components';
 import { useState } from 'react';
 import { View } from 'react-native';
-
-const valueAsSelahTeamEnum = Object.fromEntries(
-  Object.entries(enumDisplaySelahTeam).map(([key, value]) => [value, key])
-);
 
 interface ITeamProps {
   team?: SelahTeamEnum | null;
@@ -20,45 +16,52 @@ interface ITeamProps {
 
 export default function Team(props: ITeamProps) {
   const { team, noteId } = props;
-  const [localTeam, setLocalTeam] = useState<string | null | undefined>(
-    team ? enumDisplaySelahTeam[team] : null
+  const [localTeam, setLocalTeam] = useState<SelahTeamEnum | undefined | null>(
+    team
   );
 
   const [updateNote] = useUpdateNoteMutation();
   const { showSnackbar } = useSnackbar();
 
-  const updateNoteFunction = async (value: string) => {
-    if (!noteId || !value) return;
-    setLocalTeam(value);
+  const updateNoteFunction = async (selectedTeam: SelahTeamEnum) => {
+    if (!noteId || !selectedTeam) {
+      return;
+    }
+
+    setLocalTeam(selectedTeam);
 
     try {
       await updateNote({
         variables: {
           data: {
             id: noteId,
-            team: valueAsSelahTeamEnum[value] as SelahTeamEnum,
+            team: selectedTeam,
           },
         },
       });
     } catch (err) {
+      setLocalTeam(team || undefined);
+
       showSnackbar({
         message: 'Failed to update interaction.',
         type: 'error',
       });
+
       console.error('Failed to update interaction:', err);
     }
   };
 
   return (
     <View style={{ marginBottom: Spacings.xs }}>
-      <Picker
+      <SingleSelect
         placeholder="Team"
-        value={localTeam}
-        items={Object.values(SelahTeamEnum).map((item) => ({
-          label: enumDisplaySelahTeam[item],
-          value: enumDisplaySelahTeam[item],
-        }))}
-        setSelectedValue={(t) => updateNoteFunction(t as SelahTeamEnum)}
+        selectedValue={localTeam || undefined}
+        items={Object.entries(enumDisplaySelahTeam).map(
+          ([value, displayValue]) => ({ value, displayValue })
+        )}
+        onChange={(value) => {
+          updateNoteFunction(value as SelahTeamEnum);
+        }}
       />
     </View>
   );
