--- libs/expo/shared/ui-components/src/lib/Picker_V2/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Picker';
